+++ CHANGES.md
@@ -4,6 +4,8 @@
   `switch_valve_when_done` to match `wait_until_done` wording
 * Add `switch_valve_when_finished` keyword argument to
   `QmixPump.aspirate()` method
+* Wait until the pumps have actually started operating before checking
+  whether pumping has finished when using the `wait_until_done` kwarg.
 
 2018-09-13
 ----------